Move printer-specifics into separate files
